+++ src/cljs/athens/subs.cljs
@@ -1,5 +1,6 @@
 (ns athens.subs
   (:require
+    [athens.util :as util]
     [day8.re-frame.tracing :refer-macros [fn-traced]]
     [re-frame.core :as re-frame :refer [subscribe]]))
 
@@ -134,7 +135,11 @@
     (:modal db)))
 
 
+;; really bad that we're checking if electron in a subscription, but short-term solution to get both web app and desktop to build. see athens.electron
 (re-frame/reg-sub
   :db/remote-graph-conf
-  (fn [_ _]
-    nil))
+  (fn [db _]
+    (if (util/electron?)
+      (:db/remote-graph-conf db)
+      {})))
+